Refactor GV
