+++ gpm/__init__.py
@@ -34,8 +34,7 @@ from importlib.metadata import PackageNotFoundError, version
 import pycolorbar
 from pycolorbar import get_plot_kwargs  # noqa
 
-import gpm.accessor
-import gpm.xradar  # noqa
+import gpm.accessor  # noqa
 from gpm._config import config  # noqa
 from gpm.configs import (  # noqa
     define_configs,
