+++ gpm/utils/dataframe.py
@@ -70,7 +70,7 @@ def to_pandas_dataframe(ds):
 
 
 def to_dask_dataframe(ds):
-    """Convert an xarray.Dataset to a :py:class:`dask.dataframe.Dataframe`."""
+    """Convert an xarray.Dataset to a :py:class:`dask.dataframe.DataFrame`."""
     # Drop unrelevant coordinates
     ds = remove_unused_var_dims(ds)
