+++ gpm/visualization/__init__.py
@@ -27,6 +27,8 @@
 """This directory contains the GPM-API data visualization tools."""
 from gpm.visualization.plot import (
     add_map_inset,
+    plot_cartopy_background,
+    plot_colorbar,
     plot_image,
     plot_labels,
     plot_map,
@@ -41,4 +43,6 @@ __all__ = [
     "plot_image",
     "plot_map_mesh",
     "add_map_inset",
+    "plot_colorbar",
+    "plot_cartopy_background",
 ]
